Change the type of "total*" metrics
DBStats exposes metrics whose definition imply it being a counter. For
example, "wait_count" is defined as "total number of connections
waited for" which means that during the process run the value emitted by
this stat will keep increasing. Hence, it should be registered as a
"Count" metric instead of Gauge which implies a fixed value at a point
in time unrelated to previous values.